Tidy up styles into single file
