--- packages/puff/styles/index.web.js
@@ -1,8 +1,0 @@
-import { StyleSheet } from "react-native";
-import sharedStyles from "./shared";
-
-const webStyles = StyleSheet.create({
-  ...sharedStyles
-});
-
-export default webStyles;
